requirements.txt
# Django y DRF
Django==5.0
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Base de datos
psycopg2-binary==2.9.9

# Redis y Celery
redis==5.0.1
celery==5.3.4

# Utilidades
python-decouple==3.8
gunicorn==21.2.0

# Desarrollo
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Documentación
drf-spectacular==0.26.5


